Properly save and restore fearedmon property.
u.fearedmon is a pointer and was not being properly saved and
restored when saving/restoring a game. because of this, when
the `u` field is restored u.fearedmon is set the memory
address of the object of the previous lifetime of the program
which is obviously an invalid pointer most of the time on
restore. `u.usteed` and `u.ustuck` are also pointers to monst
structs and are handled properly on save/restore, so this
commit does the same thing for fearedmon.